requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-dotenv
google-generativeai
openai
reportlab
pytest
httpx
pyogrio
geopandas
